requirements.txt
-i https://pypi.org/simple
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1; python_version >= '3.7'
cryptography==44.0.1; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
google-api-core==2.24.1; python_version >= '3.7'
google-api-python-client==2.162.0; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1; python_version >= '3.6'
googleapis-common-protos==1.68.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10; python_version >= '3.6'
numpy==2.2.3; python_version < '3.11'
oauthlib==3.2.2; python_version >= '3.6'
pandas==2.2.3; python_version >= '3.9'
pdfminer.six==20231228; python_version >= '3.6'
pdfplumber==0.11.5; python_version >= '3.8'
pillow==11.1.0; python_version >= '3.9'
proto-plus==1.26.0; python_version >= '3.7'
protobuf==5.29.3; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyparsing==3.2.1; python_version >= '3.1'
pypdfium2==4.30.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.1
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2025.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
